Move WeakBox to it's own file
